refactor: MemberDto의 디렉토리를 변경하라
