+++ src/main/java/prgrms/neoike/controller/MemberController.java
@@ -5,10 +5,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import prgrms.neoike.controller.dto.MemberDto;
+import prgrms.neoike.service.dto.memberdto.MemberDto;
 import prgrms.neoike.controller.dto.MemberRequest;
 import prgrms.neoike.controller.mapper.MemberMapper;
-import prgrms.neoike.domain.member.Member;
 import prgrms.neoike.service.MemberService;
 
 import javax.validation.Valid;
@@ -21,7 +20,6 @@ public class MemberController {
 
     @PostMapping("/api/v1/members")
     public ResponseEntity<Long> joinMember(@Valid @RequestBody MemberRequest memberRequest) {
-
         MemberDto memberDto = MemberMapper.mapMemberDto(memberRequest);
         Long joinMemberId = memberService.join(memberDto);
         return ResponseEntity.ok(joinMemberId);
